fix: Use custom transparent assets instead of bevy built-in
This fixes the disappearing icon issue that happens when the icon is
just out of camera bounds (the cunning that happens). The issue is that
you could still see half of the icon, so it just pops out/in.

The reason this happens with the bevy built in transparent texture is
because we only have a weak handle and not a strong handle (I assume anyways).
It doesn't matter that we even update it to a strong handle later, it
still gets culled for some reason.